# 1_Object_Oriented/src/Presentation/Alexandria.WebApi/OpenAPI/Alexandria.WebApi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Alexandria.WebApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost"
    }
  ],
  "paths": {
    "/v1/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get a Book.",
        "operationId": "GetBook",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBookResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Add a Book.",
        "operationId": "PostBook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddBookResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookAlreadyExistsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/books/bookAuthors": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get a Book and its author(s).",
        "operationId": "GetBookAuthors",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetBookAuthorsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1/authors": {
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Get an Author.",
        "operationId": "GetAuthor",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAuthorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Authors"
        ],
        "summary": "Add an Author.",
        "operationId": "PostAuthor",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAuthorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddAuthorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorAlreadyExistsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/accounts": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Get an Account.",
        "operationId": "GetAccount",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAccountResponse"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddAuthorRequest": {
        "required": [
          "firstName",
          "middleNames",
          "lastName",
          "birthDate"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "middleNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AddAuthorResponse": {
        "required": [
          "author"
        ],
        "type": "object",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/Author"
          }
        }
      },
      "AddBookRequest": {
        "required": [
          "title",
          "publicationDate",
          "authorsIds"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "publicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "authorsIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "AddBookResponse": {
        "required": [
          "book"
        ],
        "type": "object",
        "properties": {
          "book": {
            "$ref": "#/components/schemas/Book"
          }
        }
      },
      "Author": {
        "required": [
          "id",
          "firstName",
          "middleNames",
          "lastName",
          "birthDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "middleNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuthorAlreadyExistsResponse": {
        "required": [
          "author"
        ],
        "type": "object",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Book": {
        "required": [
          "id",
          "title",
          "publication",
          "authorsIds"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "publication": {
            "$ref": "#/components/schemas/Publication"
          },
          "authorsIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
      "BookAlreadyExistsResponse": {
        "required": [
          "book"
        ],
        "type": "object",
        "properties": {
          "book": {
            "$ref": "#/components/schemas/Book"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "GetAccountResponse": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          }
        }
      },
      "GetAuthorResponse": {
        "required": [
          "author"
        ],
        "type": "object",
        "properties": {
          "author": {
            "$ref": "#/components/schemas/Author"
          }
        }
      },
      "GetBookAuthorsResponse": {
        "required": [
          "book",
          "authors"
        ],
        "type": "object",
        "properties": {
          "book": {
            "$ref": "#/components/schemas/Book"
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Author"
            }
          }
        }
      },
      "GetBookResponse": {
        "required": [
          "book"
        ],
        "type": "object",
        "properties": {
          "book": {
            "$ref": "#/components/schemas/Book"
          }
        }
      },
      "Publication": {
        "required": [
          "id",
          "publicationDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "publicationDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Books"
    },
    {
      "name": "Authors"
    },
    {
      "name": "Accounts"
    }
  ]
}